Add TaskAuth class with verifyTaskOwnership() to taskController methods

+++ src/controller/TaskController.js
@@ -13,8 +13,13 @@ export class TaskController {
 
     static async get(req, res) {
         const { id } = req.params
+        const { id: USER_ID } = req.user
+        
         try {
+            await TaskAuth.verifyTaskOwnership(id, USER_ID)
+            
             const task = await Task.getTaskById(id)
+            
             if (task) {
                 res.json(task)
             } else {
@@ -28,8 +33,13 @@ export class TaskController {
     static async update(req, res) {
         const { id } = req.params
         const { description_task, status_task, due_date } = req.body
+        const { id: USER_ID } = req.user
+
         try {
+            await TaskAuth.verifyTaskOwnership(id, USER_ID)
+
             await Task.updateTask(id, description_task, status_task, due_date)
+            
             res.json({ message: 'Task updated successfully' })
         } catch (error) {
             res.status(500).json({ error: 'Error updating task' })
@@ -38,11 +48,26 @@ export class TaskController {
 
     static async delete(req, res) {
         const { id } = req.params
+        const { id: USER_ID } = req.user
+        
         try {
+            await TaskAuth.verifyTaskOwnership(id, USER_ID)
+
             await Task.deleteTask(id)
+            
             res.json({ message: 'Task deleted successfully' })
         } catch (error) {
             res.status(500).json({ error: 'Error deleting task' })
         }
+    }
+}
+
+export class TaskAuth {
+    static async verifyTaskOwnership(task_id, USER_ID) {
+        const TARGETED_TASK = await Task.getTaskById(task_id)
+            
+        if (!TARGETED_TASK) throw new Error('Task not found')
+
+        if (TARGETED_TASK.id_user !== USER_ID) throw new Error('Invalid task ID')
     }
 }
